multiagent02/pyproject.toml
[project]
name = "multiagent02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.3.0",
    "openai>=0.27.10",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
